Remove codeduplication and move the interface in its on file
